+++ bos/core/inc/b_core.h
@@ -108,8 +108,8 @@ typedef struct
  */
 
 int bOpen(uint8_t dev_no, uint8_t flag);
-int bRead(int fd, uint8_t *pdata, uint16_t len);
-int bWrite(int fd, uint8_t *pdata, uint16_t len);
+int bRead(int fd, uint8_t *pdata, uint32_t len);
+int bWrite(int fd, uint8_t *pdata, uint32_t len);
 int bCtl(int fd, uint8_t cmd, void *param);
 int bLseek(int fd, uint32_t off);
 int bClose(int fd);
