+++ bos/drivers/b_drv_lis3dh.c
@@ -36,7 +36,6 @@
 
 #include "utils/inc/b_util_log.h"
 
-
 /**
  * \addtogroup B_DRIVER
  * \{
@@ -318,10 +317,10 @@ static int _bLis3dhCtl(bLIS3DH_Driver_t *pdrv, uint8_t cmd, void *param)
     return retval;
 }
 
-static int _bLis3dhRead(bLIS3DH_Driver_t *pdrv, uint32_t off, uint8_t *pbuf, uint16_t len)
+static int _bLis3dhRead(bLIS3DH_Driver_t *pdrv, uint32_t off, uint8_t *pbuf, uint32_t len)
 {
     bLis3dhFifoSrcReg_t fifo_src_reg;
-    bGsensor3Axis_t *   pTemp = (bGsensor3Axis_t *)pbuf;
+    bGsensor3Axis_t    *pTemp = (bGsensor3Axis_t *)pbuf;
     int                 i     = 0;
     uint8_t             fss   = 0;
     uint8_t             c     = len / sizeof(bGsensor3Axis_t);
