+++ bos/drivers/b_drv_oled.c
@@ -247,7 +247,7 @@ static int _bOLEDCtl(bOLED_Driver_t *pdrv, uint8_t cmd, void *param)
     return retval;
 }
 
-static int _bOLEDWrite(bOLED_Driver_t *pdrv, uint32_t addr, uint8_t *pbuf, uint16_t len)
+static int _bOLEDWrite(bOLED_Driver_t *pdrv, uint32_t addr, uint8_t *pbuf, uint32_t len)
 {
     uint16_t     x      = addr % LCD_X_SIZE;
     uint16_t     y      = addr / LCD_X_SIZE;
