+++ bos/drivers/b_drv_ssd1289.c
@@ -208,7 +208,7 @@ static int _bSSD1289Ctl(bSSD1289_Driver_t *pdrv, uint8_t cmd, void *param)
     return retval;
 }
 
-static int _bSSD1289Write(bSSD1289_Driver_t *pdrv, uint32_t addr, uint8_t *pbuf, uint16_t len)
+static int _bSSD1289Write(bSSD1289_Driver_t *pdrv, uint32_t addr, uint8_t *pbuf, uint32_t len)
 {
     uint16_t     x      = addr % LCD_X_SIZE;
     uint16_t     y      = addr / LCD_X_SIZE;
