+++ bos/drivers/b_drv_st7789.c
@@ -277,7 +277,7 @@ static int _bST7789Ctl(bST7789_Driver_t *pdrv, uint8_t cmd, void *param)
     return retval;
 }
 
-static int _bST7789Write(bST7789_Driver_t *pdrv, uint32_t addr, uint8_t *pbuf, uint16_t len)
+static int _bST7789Write(bST7789_Driver_t *pdrv, uint32_t addr, uint8_t *pbuf, uint32_t len)
 {
     uint16_t     x      = addr % LCD_X_SIZE;
     uint16_t     y      = addr / LCD_X_SIZE;
