+++ bos/mcu/hdsc/hc32l13x/mcu_hc32l13x_flash.c
@@ -158,11 +158,11 @@ int bMcuFlashErase(uint32_t raddr, uint8_t pages)
     return retval;
 }
 
-int bMcuFlashWrite(uint32_t raddr, const uint8_t *pbuf, uint16_t len)
+int bMcuFlashWrite(uint32_t raddr, const uint8_t *pbuf, uint32_t len)
 {
-    int     timeout = 0;
-    uint8_t wdata   = 0;
-    uint8_t wlen = len, i = 0;
+    int      timeout = 0;
+    uint8_t  wdata   = 0;
+    uint32_t wlen = len, i = 0;
     raddr = FLASH_BASE_ADDR + raddr;
     if (pbuf == NULL || (raddr & 0x1) || (raddr + len) > (FLASH_MAX_SIZE + FLASH_BASE_ADDR) ||
         ((MCU_FLASH->CR) & (0x1 << 4)) != 0)
@@ -225,7 +225,7 @@ int bMcuFlashWrite(uint32_t raddr, const uint8_t *pbuf, uint16_t len)
     return wlen;
 }
 
-int bMcuFlashRead(uint32_t raddr, uint8_t *pbuf, uint16_t len)
+int bMcuFlashRead(uint32_t raddr, uint8_t *pbuf, uint32_t len)
 {
     if (pbuf == NULL || (raddr + FLASH_BASE_ADDR + len) > (FLASH_MAX_SIZE + FLASH_BASE_ADDR))
     {
